Add option to rewrite binary
Here we introduce the -rewrite option and change section-based address
assignment to segment-based. We handle more relocations, and determine
the final layout by iterating over segments, determining which sections
they contain and either copying them(old logic) or assigning new
addresses(-rewrite). Instead of ugly patchELFPHDRTable we have a vector of
output segments which is populated by mapping functions and then written
to file in a simple loop.
We introduce getOutputFileOffsetForAddress and outputMappings map to
decouple input from output layout and use it in a bunch of places
instead of old getFileOffsetForAddress.
We also have to determine size of .eh_frame_hdr in advance, so we're
parsing .eh_frame in mapEhFrames to find out how much entries it countains
and compute frame header size based on that.
getOutputSections becomes more complicated to properly handle index
changes, but i think it can be improved in the future.