Add AArch64 PLT emission and handle GOT references
We emit AArch64 PLT as functions and use GOTSymbolsByName map to properly handle
adrp+ldr. The map is populated with the help of dynamic relocations in .got or
the values contained there. Unfortunately matching X86 .plt entries is not as easy
as in AArch64, so for X86 we continue emitting PLT as data.