Refactor section handling
EFMM after https://reviews.llvm.org/D135494 managed section names in the
following manner: if a section was emitted with orgSecPrefix, it updated
original BinarySection, and if it was emitted with original name, the
section was duplicated with newSecPrefix. The problem is that we can't
control what name is used to emit .text or .eh_frame, and the old code
made it impossible to rewrite these sections in-place. Here we remove
all the management from EFMM and explicitly rename sections in
RewriteInstance if we want them duplicated, or don't touch them when we
rewrite, thus allowing EFMM to update them. We also have to rename some
sections like eh_frame back in case they were not emitted, which is a
bit combersome, but overall we get more flexibility over what sections
we want and under which names. In case of -reorder-data pass, i think
the simplest solution is to just create new section with .hot suffix,
and emit it under this name.