[BOLT] Improve handling of end-of-section symbols
There are at least two issues with such symbols. Firsly, they don't fit
into BOLT's notion of BinaryData. We cannot register them as regular
data because those data objects are identified by addresses, and
registering multiple symbols at the same address implies that it is the same
entity with alternative names which will have the same output address.
It is not the case with end symbols, since they're usually located on
the border between two sections. So, calling registerNameAtAddress with
end symbol will very likely register the name as alternative for some
function or data from the next section and result in incorrect output
address. Because of this, we have to register end-of-section symbols in
a separate map and check every symbol name against it while resolving
relocations, updating symtab and patching dynamic relocations.

And patching dynamic relocations is the second issue. A lot of dynamic
relocations are R_*_RELATIVE and only contain address without symbol
reference. So, to disambiguate the end of one section from the beginning
of another, we have to look at static relocation on the same offset,
which fortunately contains the symbol reference which we
can use to get symbol name and look up if it is end of section marker.

However, this activity interferes a bit with BOLT moving and
copying sections, and extracting some data into new ones (Golang pass),
so we have to be careful and it's probable that even with all those
lookups into EndSymbols map we'll have to special case some more stuff.