[BOLT] Put .plt before .text when rewriting
Having PLT in arbitrary location caused issues for LongJMP in AArch64,
because we don't know the final binary layout at LongJMP pass and thus
cannot decide when to insert stubs for PLT calls. Previously the code just assumed
that PLT address will not change, which caused assertion failures for big
binaries or when -instrument was combined with -rewrite and .plt was moved
by inflated text section. While it is not possible to know the binary layout
before linking, we can just forcefully put PLT before .text which will allow us to
easily estimate it's position relative to .text and thus properly insert stubs.
As a bonus, we get more locality because we put .plt near hot text.