[BOLT] Exit when --rewrite passed without relocations
There is some code in RI::discoverStorage which makes decisions
on binary layout depending on whether --rewrite is set. Once they
are made, we can't change the mode to non-rewrite if relocations
are not present, because for non-rewrite case we had to compute
new .text address in advance in discoverStorage, and we can only
check for .rela sections after discoverStorage is finished.